Added first tests and restructured project in a more modular fashion
