+++ .eslintrc
@@ -1,3 +1,6 @@
 {
-  "extends": "standard"
+  "extends": "standard",
+  "env": {
+    "jest": true
+  }
 }
