+++ package.json
@@ -11,10 +11,12 @@
     "url": "https://raw.githubusercontent.com/lmammino/godaddy-dns/master/LICENSE"
   },
   "bin": {
-    "godaddy-dns": "./godaddy-dns.js"
+    "godaddy-dns": "src/cli.js"
   },
   "scripts": {
-    "test": "eslint godaddy-dns.js",
+    "test:lint": "eslint godaddy-dns.js",
+    "test:unit": "jest src --verbose --coverage",
+    "test": "npm run test:lint && npm run test:unit",
     "package:create": "pkg . -t node6-win,node6-macos,node6-linux --out-dir build",
     "package:publish": "node scripts/publish-release $GITHUB_TOKEN lmammino/godaddy-dns $npm_package_version",
     "release:tag": "git tag $npm_package_version && git push --tags"
@@ -47,6 +49,7 @@
     "eslint-plugin-node": "^5.2.1",
     "eslint-plugin-promise": "^3.6.0",
     "eslint-plugin-standard": "^3.0.1",
+    "jest": "^21.2.1",
     "pkg": "^4.3.0-beta.1"
   }
 }
