+++ src/getCurrentIp.js
@@ -1,0 +1,5 @@
+const request = require('request-promise')
+
+module.exports = function getCurrentIp () {
+  return request('https://api.ipify.org/')
+}
